requirements.txt
Django==4.1.3
django-ckeditor-5==0.2.18
django-cleanup==8.1.0
django-compressor==4.5
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-taggit==5.0.1
django-markdown-deux==1.0.6
Markdown==3.5.2
markdown2==2.4.12
Pillow==11.0.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1